# docs/企业级功能完善与统一性设计方案/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: ZKER Enterprise API
  description: |
    ZKER Enterprise 企业级 Multi-Tenant SaaS AI 平台 API 文档

    ## 功能概述
    - 🤖 **AI对话管理**: 智能对话、多轮交互、插件调用
    - 📊 **数据分析**: NL2SQL、数据可视化、智能分析
    - ✍️ **AI写作**: AI写作、模板生成、多风格
    - 🏢 **企业管理**: 组织架构、权限管理、数字员工
    - 🔧 **开发平台**: Bot构建、工作流编辑、插件开发
    - 📈 **监控运维**: 性能监控、日志分析、告警管理

    ## 认证方式
    API 使用 Bearer Token 进行认证：
    ```
    Authorization: Bearer {token}
    ```

    ## 租户识别
    支持三种租户识别方式：
    1. **Subdomain**: `{tenant}.zker.com`
    2. **Path**: `/api/v1/{tenant}/...`
    3. **Header**: `X-Tenant-ID: {tenant_id}`

  version: 1.0.0
  contact:
    name: ZKER Enterprise Team
    email: api@zker.com
  license:
    name: Proprietary
servers:
  - url: https://api.zker.com/v1
    description: 生产环境
  - url: https://api-staging.zker.com/v1
    description: 预发布环境
  - url: http://localhost:8080/v1
    description: 本地开发环境
...
paths:
...
  /bots:
    get:
      tags:
        - bots
      summary: 获取Bot列表
      description: 获取当前用户可访问的Bot列表
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, archived]
          description: 按状态筛选
        - name: search
          in: query
          schema:
            type: string
          description: 搜索关键词
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 0
                  message:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Bot'
                      total:
                        type: integer
                        example: 100
                      page:
                        type: integer
                        example: 1
                      page_size:
                        type: integer
                        example: 20

    post:
      tags:
        - bots
      summary: 创建Bot
      description: 创建新的AI Bot
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BotCreateRequest'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 0
                  message:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/Bot'
        '400':
          $ref: '#/components/responses/BadRequest'
  /bots/{bot_id}:
    get:
      tags:
        - bots
      summary: 获取Bot详情
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: bot_id
          in: path
          required: true
          schema:
            type: string
          description: Bot ID
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 0
                  message:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/BotDetail'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - bots
      summary: 更新Bot
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: bot_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BotUpdateRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 0
                  message:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/Bot'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - bots
      summary: 删除Bot
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: bot_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /bots/{bot_id}/publish:
    post:
      tags:
        - bots
      summary: 发布Bot
      description: 将Bot发布到指定渠道
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: bot_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - channels
              properties:
                channels:
                  type: array
                  items:
                    type: string
                    enum: [web, wechat, feishu, discord]
                  example: ["web", "wechat"]
                config:
                  type: object
                  description: 渠道配置信息
                  properties:
                    wechat_app_id:
                      type: string
                    wechat_app_secret:
                      type: string
      responses:
        '200':
          description: 发布成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # ==================== 对话管理 ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token认证，格式：Bearer {token}
  parameters:
    TenantId:
      name: X-Tenant-ID
      in: header
      description: 租户ID
      required: false
      schema:
        type: string
      example: tenant_001
    Page:
      name: page
      in: query
      description: 页码
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      description: 每页数量
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        code:
          type: integer
          description: 响应码，0表示成功
          example: 0
        message:
          type: string
          description: 响应消息
          example: success
        data:
          type: object
          description: 响应数据
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          description: 错误码
          example: 10001
        message:
          type: string
          description: 错误消息
          example: 参数错误
        details:
          type: object
          description: 错误详情

    # ==================== 认证相关 ====================
...
    Bot:
      type: object
      properties:
        id:
          type: string
          example: bot_123
        name:
          type: string
          example: "客服助手"
        description:
          type: string
          example: "智能客服Bot"
        avatar:
          type: string
          format: uri
          example: "https://cdn.zker.com/avatars/bot_123.png"
        status:
          type: string
          enum: [active, inactive, archived]
          example: active
        type:
          type: string
          enum: [chat, workflow, agent]
          example: chat
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    BotDetail:
      allOf:
        - type: object
          properties:
            id:
              type: string
            name:
              type: string
            description:
              type: string
            avatar:
              type: string
            status:
              type: string
            type:
              type: string
        - type: object
          properties:
            prompt_template:
              type: string
              description: 提示词模板
            model_config:
              type: object
              description: 模型配置
              properties:
                model:
                  type: string
                  example: "gpt-4"
                temperature:
                  type: number
                  minimum: 0
                  maximum: 2
                  example: 0.7
                max_tokens:
                  type: integer
                  example: 2000
            plugins:
              type: array
              items:
                $ref: '#/components/schemas/Plugin'
            knowledge_bases:
              type: array
              items:
                $ref: '#/components/schemas/KnowledgeBase'
            statistics:
              $ref: '#/components/schemas/BotStatistics'
    BotCreateRequest:
      type: object
      required:
        - name
        - type
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 500
        type:
          type: string
          enum: [chat, workflow, agent]
        prompt_template:
          type: string
        model_config:
          type: object
          properties:
            model:
              type: string
            temperature:
              type: number
            max_tokens:
              type: integer
    BotUpdateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
          maxLength: 500
        status:
          type: string
          enum: [active, inactive, archived]
        prompt_template:
          type: string
        model_config:
          type: object
    BotStatistics:
      type: object
      properties:
        total_conversations:
          type: integer
          description: 总对话数
          example: 10000
        total_messages:
          type: integer
          description: 总消息数
          example: 50000
        active_users:
          type: integer
          description: 活跃用户数
          example: 500
        avg_response_time:
          type: number
          description: 平均响应时间(ms)
          example: 1200
        satisfaction_score:
          type: number
          description: 满意度评分(0-5)
          example: 4.5
...
    KnowledgeBase:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [document, qa, dataset]
        document_count:
          type: integer
          description: 文档数量
        chunk_count:
          type: integer
          description: 分块数量
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # ==================== 插件相关 ====================
    Plugin:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [search, productivity, entertainment, tools]
        icon:
          type: string
          format: uri
        version:
          type: string
          example: "1.0.0"
        author:
          type: string
        rating:
          type: number
          minimum: 0
          maximum: 5
        install_count:
          type: integer
        is_installed:
          type: boolean

    # ==================== 用户相关 ====================
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 400
            message: 参数错误
            details:
              username: 用户名不能为空
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 404
            message: 资源不存在
...